backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.12.0
aiofiles==23.2.1
reportlab==4.0.8
openpyxl==3.1.2
